group info and param tests
