+++ test_autofit/mapper/conftest.py
@@ -1,0 +1,8 @@
+import pytest
+
+import autofit as af
+
+
+@pytest.fixture(name="mapper")
+def make_mapper():
+    return af.ModelMapper()
